feat: Implement employee allowance assignment functionality
In this commit, we have added the following changes:
- Created a new form for assigning allowances to employees.
- Implemented the logic to handle allowance assignments.
- Added a "Back" button to return to the employee's allowances page.

Closes #<issue-number>

+++ HRMSZPLUS/src/main/java/com/HRMS/service/EmployeeService.java
@@ -8,7 +8,7 @@ public interface EmployeeService {
 	
 	List<EmployeeMaster> getAllEmployees();
 	EmployeeMaster addEmployee(EmployeeMaster employee);
-	EmployeeMaster findById(int id);
-	boolean deleteEmployee(int id);
+	EmployeeMaster findById(long id);
+	boolean deleteEmployee(long id);
 
 }
